+++ pysat/_instrument.py
@@ -109,7 +109,9 @@ class Instrument(object):
     yr : int
         year for loaded data
     bounds : (datetime/filename/None, datetime/filename/None)
-        bounds for loading data, supply array_like for a season with gaps
+        bounds for loading data, supply array_like for a season with gaps.
+        Users may provide as a tuple or tuple of lists, but the attribute is
+        stored as a tuple of lists for consistency
     doy : int
         day of year for loaded data
     files : pysat.Files
@@ -443,7 +445,10 @@ class Instrument(object):
                 try:
                     return self.data.isel(time=key[0])[key[1]]
                 except:
-                    return self.data.sel(time=key[0])[key[1]]
+                    try:
+                        return self.data.sel(time=key[0])[key[1]]
+                    except TypeError: # construct dataset from names
+                        return self.data[self.variables[key[1]]]
             else:
                 # multidimensional indexing
                 indict = {}
@@ -698,12 +703,44 @@ class Instrument(object):
         return copy.deepcopy(self)
 
     def concat_data(self, data, *args, **kwargs):
-        """Concats data1 and data2 for xarray or pandas as needed"""
+        """Concats data1 and data2 for xarray or pandas as needed
+
+        Note
+        ----
+        For pandas, sort=False is passed along to the underlying
+        pandas.concat method. If sort is supplied as a keyword, the
+        user provided value is used instead.
+
+        For xarray, dim='time' is passed along to xarray.concat
+        except if the user includes a value for dim as a
+        keyword argument.
+
+        Parameters
+        ----------
+        data : pandas or xarray
+           Data to be appended to data already within the Instrument object
+
+        Returns
+        -------
+        void
+            Instrument.data modified in place.
+
+        """
 
         if self.pandas_format:
-            return pds.concat(data, sort=True, *args, **kwargs)
+            if 'sort' in kwargs:
+                sort = kwargs['sort']
+                _ = kwargs.pop('sort')
+            else:
+                sort = False
+            return pds.concat(data, sort=sort, *args, **kwargs)
         else:
-            return xr.concat(data, dim='time')
+            if 'dim' in kwargs:
+                dim = kwargs['dim']
+                _ = kwargs.pop('dim')
+            else:
+                dim = 'time'
+            return xr.concat(data, dim=dim, *args, **kwargs)
 
     def _pass_func(*args, **kwargs):
         pass
@@ -2023,8 +2060,8 @@ class Instrument(object):
         All attributes attached to instrument meta are written to netCDF attrs
         with the exception of 'Date_End', 'Date_Start', 'File', 'File_Date',
         'Generation_Date', and 'Logical_File_ID'. These are defined within to_netCDF
-        at the time the file is written, as per the adopted standard, 
-        SPDF ISTP/IACG Modified for NetCDF. Atrributes 'Conventions' and 
+        at the time the file is written, as per the adopted standard,
+        SPDF ISTP/IACG Modified for NetCDF. Atrributes 'Conventions' and
         'Text_Supplement' are given default values if not present.
 
         """
@@ -2450,7 +2487,7 @@ class Instrument(object):
             if 'Conventions' not in adict:
                 adict['Conventions'] = 'SPDF ISTP/IACG Modified for NetCDF'
             if 'Text_Supplement' not in adict:
-                adict['Text_Supplement'] = ''                
+                adict['Text_Supplement'] = ''
             # remove any attributes with the names below
             # pysat is responible for including them in the file.
             items = ['Date_End', 'Date_Start', 'File', 'File_Date',
@@ -2464,7 +2501,7 @@ class Instrument(object):
                                         '%a, %d %b %Y,  ' +
                                         '%Y-%m-%dT%H:%M:%S.%f')
             adict['Date_End'] = adict['Date_End'][:-3] + ' UTC'
-            
+
             adict['Date_Start'] = \
                 pysat.datetime.strftime(self.index[0],
                                         '%a, %d %b %Y,  ' +